Refactor: move the persisted settings shape into lib
`plan-share-codec.ts` imported from `@/hooks/useDomainSettings` — the
only lib→hooks import in the codebase, and one this branch introduced.
The cause was that the persistence layer lived inside the hook: the
shape type, its migration, its validation and the compose/persist pair
are all pure, but three consumers need them and only one is a hook.

They move to `src/lib/persisted-shape.ts`, leaving `useDomainSettings`
with React state and actions (1731 -> 1111 lines).

That makes `sanitizePersistedShape` directly testable, which matters
more than the layering: it is the single validation path for every
settings payload, including URL content, and it was only ever exercised
through the codec — i.e. with input the encoder produced, which is
exactly the input that is never dangerous. The new suite feeds it
hostile payloads directly and pins the contract: unknown ids dropped
rather than admitted, non-finite and negative numbers rejected, and the
`currentDomain ∈ activeDomains` invariant (which App.tsx non-null-asserts
on) restored on the way out.

Adds a completeness guard for `DELTA_KEYS`. `satisfies` only rejected
keys that aren't on `PersistedShape`; the dangerous direction was the
other one — add a settings category, forget the list, and every shared
link silently omits it while the receiver computes a different plan.
The guard fails the build and names the missing field.

Also retargets a hostile-input test that claimed `0` was a valid item
code. It isn't — index 0 is a tombstone — so the test passed through the
unknown-id path and duplicated its neighbour, leaving the case it
described (a real item that isn't a raw of that region) uncovered.
